FWMT-554 changing around of config classes added a shared config created a builder class for connection factories
