+++ src/main/java/uk/gov/ons/fwmt/fwmtrmadapter/config/RMQueueConfig.java
@@ -1,11 +1,11 @@
 package uk.gov.ons.fwmt.fwmtrmadapter.config;
 
+import org.aopalliance.aop.Advice;
 import org.springframework.amqp.core.Binding;
 import org.springframework.amqp.core.BindingBuilder;
 import org.springframework.amqp.core.DirectExchange;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.core.QueueBuilder;
-import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.amqp.rabbit.listener.SimpleMessageListenerContainer;
 import org.springframework.amqp.rabbit.listener.adapter.MessageListenerAdapter;
@@ -13,7 +13,10 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.retry.interceptor.RetryOperationsInterceptor;
 import uk.gov.ons.fwmt.fwmtrmadapter.message.impl.RMReceiverImpl;
+
+import static uk.gov.ons.fwmt.fwmtrmadapter.config.ConnectionFactoryBuilder.createConnectionFactory;
 
 @Configuration
 public class RMQueueConfig {
@@ -82,9 +85,13 @@ public class RMQueueConfig {
   @Bean
   public SimpleMessageListenerContainer rmContainer(
       @Qualifier("rmConnectionFactory") ConnectionFactory connectionFactory,
-      @Qualifier("rmListenerAdapter") MessageListenerAdapter messageListenerAdapter) {
+      @Qualifier("rmListenerAdapter") MessageListenerAdapter messageListenerAdapter,
+      @Qualifier("interceptor") RetryOperationsInterceptor retryOperationsInterceptor) {
     SimpleMessageListenerContainer container = new SimpleMessageListenerContainer();
 
+    Advice[] adviceChain = {retryOperationsInterceptor};
+
+    container.setAdviceChain(adviceChain);
     container.setConnectionFactory(connectionFactory);
     container.setQueueNames(ACTION_FIELD_QUEUE);
     container.setMessageListener(messageListenerAdapter);
@@ -94,14 +101,6 @@ public class RMQueueConfig {
   // Connection Factory
   @Bean
   public ConnectionFactory rmConnectionFactory() {
-    CachingConnectionFactory cachingConnectionFactory = new CachingConnectionFactory();
-
-    cachingConnectionFactory.setPort(rmPort);
-    cachingConnectionFactory.setHost(hostname);
-    cachingConnectionFactory.setVirtualHost(virtualHost);
-    cachingConnectionFactory.setPassword(password);
-    cachingConnectionFactory.setUsername(username);
-
-    return cachingConnectionFactory;
+    return createConnectionFactory(rmPort, hostname, virtualHost, password, username);
   }
 }
